cx_clips: first version of ClipsManager and Client
This is fairly big commit, including the first
full implementation version of the ClipsEnvManager
and Client for the manager.

The Manager is mainly based on the Fawkes-ClipsEnvManager
implemented by -  Tim Niuemuller [www.niemuller.de]

The manager is implemented as a Lifecycle Node so it's
behavior can easily be controlled.
Main differences: functions to interact with the manager
are implemented as ROS2 services. Currently:
- clips_manager/create_env
- clips_manager/destroy_env
- clips_manager/add_clips_features
- clips_manager/assert_can_remove_features
- clips_manager/remove_features

The interaction with the ClipsEnvManager happens through
the ClipsEnvaManagerClient, who can be seen as wrapper
(API). It creates client to all the availabe services
from the manager.
This way, other nodes/plugins/packages can communicate with
the Manager through it's client and thus won't need to always
implement client services and/or have a local instance of it,
as the manager is fairly huge.

In order to initialise/destroy the context of the underlying
feature, the manager will communicate with a running node in
the face of the ClipsFeaturesManager, which will interact with the
registered Features.