cx_features: add FeatureManager and MockFeature
This commit includes the alpha-version implementation
of the Features manager.
It's functionallity will be to save instance of all
the registered features (later as part of laucn/params).

It will be called from the ClipsEnvManager via the two
available services to either call the clips context
initialisation/destruction of the requested feature.
The FeatureManager will then call the corresponding feature
function and send feedback to the manager.

Also a MockFeature inheriting from the ClipsFeature
interface is included, which will be then used for a
simple unit test.