cx_features: load registered features as plugins in Manager
The FeatureManager utilises the pluginlib functionallity
and loads the features, which are registered as plugins
inside the cx_features package.
Each new feature needs to inherit from the cx:ClipsFeature class
and be registered as a plugin in the cpp and added as a library in the
cmake file.
Then, each feature should be added to the cx_bringup params file with
the appropriate fields. This will ensure it's loading by the Manager.