cx_lifecycle_manager: add lifecycle nodes controller
This package will serve the role of a controller for
the running lifecycle nodes as recommended by the docs.
Each lc node gets assigned a client instance, which
will be a client for the services get_state and change_state
and implement the request to those services.

The manager will then create the corresponding clients
and start an initialisation script to configure and activate
the specified lc_nodes from the launch file.
Furthermore, the manager will implement  disable/pause/resume fncs
for the lc_nodes.