cx_features: create config feature
This features mimic the functionallity of the
fawkes config features. It depends on the yaml-cpp
library to parse and iterate through the yaml config
file. As yaml-cpp does not provide Scalar types,
I've implemented such YAML namespace functionallity.
There are also recursive functions that allow the
iteration trough the whole file, while adding
the corresponding confval to CLIPS.

The feature provides a config-load function to the
Clips Env to extract the configuration values of a
given config prefix.

All Config files will be saved under:
cx_bringup/params/*