cx_clips_executive: initial version of the clips_executive
The current implementation is fetures independent, except for
config-, and redefine_warning_feature.
The CX is a lifecycle node, which registers a clips env
executive and at start maps actions, path-substitutions and
spec name based on the clips_executive yaml file.
As it is clips-based, after activation it refreshes and runs the
clips-agenda in a timer_callback.
The main building blocks are the clips files.
The fawkes test-scenario is almost integrated in the context of
the ros cx, other than the pddl-parser and skiller.