cx_blackboard: implement InterfaceNode and Manager
- InterfaceNode will be a template based ros2 node, which will
create the corresponding interface and allow other nodes to write/
read from it using services
- InterfaceManager will provide the creation of interface nodes,
as well as setting general information for the corresponding node.

InterfaceNode is a template class, expecting the interface service
message for communication as well as the actual interface message.
The ros2 generated messages/services do not have a base class, so
dynamic polymorphism is not suitable.
Therefore, the manager holds a map of the possible interface node
types, which are saved inside an std::variant and for casting inside
std::optional.

Drawback: each interface must be included inside the manager and the
corresponding casting must be provided
-> hardcode to some degree