clips: add pddl implementation
The pddl clips file is loaded, when the executive and
the plansys2 features are available.
Clips uses the plansys2 feature to communicate with plansys2 by
adding the current domain facts and objects to the problem expert and than
calling for a plan.
Multiple plans can be requested at a time, but the planning step is
strictly sequential.