cx_skills: add skill execution mechanism
Possible to implement a given skill using the
SkillExecution class as a base class and writing
the execution logic inside perform_execution.

SkillExecutionMaster controlls the execution:
request, cancellation, confirmation. It is used
by the skill execution feature, so the clips env
can start a skill execution by providing the master
with the corresponding information about the skill.

Both Master and SkillExecution publish/are subbed to the
same topic - /skill_board and answer on different request types.