cx_nav2_example: first proof of concept using nav2 tb3 simu
implements a nav2 move skill node, which send navigate
through pose commands to the nav2 stack.
It uses the  tb3 simulation launch and the cx_with_dummy_skill
launch. The clips env will send commands to the skill master,
which will then send commands to the move action skill to navigate
to the given waypoint