cx: fix time and CLIPS loop
ROS2 uses a unix time based time format represented by a double. Fawkes
used a tuple for representation consisting of time in seconds and a
second entry for milliseconds. To keep everything consistent with ROS2,
the timing functions were changed and some mutex bugs in the main loop
were fixed.