cx_clips_executive: refactor usage of core functionalities
Let users completely decide which files of the core to use, aside from
saliences.clp and init.clp.

For clarity, the code is now put into a core subdirectory to hopefully
avoid confusion when loading files from the agent_dir.

Also, the simple_agent is now moved to the bringup package where it
belongs.