cx_ros_msgs_plugin: protect maps and default initialization
As multiple threads could be writing to the hash maps, one recursive
mutex serves as a safety mechanism.

The "recursive" part is necessary for situations, when functions such as
ros-msgs-get-field are called from a LHS of a rule, which therefore get
evaluated e.g., when the subscription callback asserts a fact in the
environment.