cx_protobuf_plugin: add option to use linked protobuf messages
protobuf message registers either work through lookup of linked messages
in the shared library or through reading proto files directly.

The latter was already supported through cx_protobuf_plugin, the former
requires to link the compiled proto definitions into the plugin, which
does not directly work.

Instead, offer a cmake macro that generates plugin based on
an existing shared library or proto sources and the cx_protobuf_plugin.

This allows to use the generated plugin instead of the cx_protobuf_plugin for
usage with lookups from the linked messages.