ros_msgs_plugin: keep shared ptr of sub-msg in parent scope
Description of old erronous behavior:
When a sub-message is added (through calling get-field on a parent
message), we add it to the list of sub-messages of the parent
(sub_messages_ map).

If we delete a sub-message before the parent message, it goes out of
scope and we did not delet the respective entries in the list of
sub-messages of the parent.

When a new message is created that gets the same address as the deleted
sub-message, the parent message will clean it up thinking it is the
sub-message.

Fix: Just store the shared pointers in the list of sub-messges so they
actually do not go out of scope until the parent message is deleted.